backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
requests
